Extract shared Gmail utilities
