+++ gmail_bot.py
@@ -3,11 +3,16 @@ import os
 # additional imports
 import json
 import requests
+import base64
 
-from googleapiclient.discovery import build     # already imported in Draft_Replies, keep here too
-from google.auth.transport.requests import Request
-from google_auth_oauthlib.flow import InstalledAppFlow
-import pickle, base64
+from gmail_utils import (
+    get_gmail_service,
+    fetch_all_unread_messages,
+    create_base64_message,
+    create_draft,
+    is_promotional_or_spam,
+    classify_email,
+)
 
 from openai import OpenAI
 from Draft_Replies import generate_ai_reply
@@ -17,10 +22,6 @@ import yaml
 with open(os.path.join(os.path.dirname(__file__), "config.yaml")) as f:
     CFG = yaml.safe_load(f)
 
-# Gmail settings
-SCOPES               = CFG["gmail"]["scopes"]
-GMAIL_CLIENT_SECRET  = CFG["gmail"]["client_secret_file"]
-GMAIL_TOKEN_FILE     = CFG["gmail"]["token_file"]
 
 # OpenAI models & API key
 OPENAI_API_KEY       = os.getenv(CFG["openai"]["api_key_env"])
@@ -50,110 +51,6 @@ FREESCOUT_KEY        = CFG["ticket"]["freescout_key"]
 
 if not OPENAI_API_KEY:
     raise ValueError(f"Please set your {CFG['openai']['api_key_env']} environment variable.")
-
-def get_gmail_service(creds_filename=None, token_filename=None):
-    """Authenticate with Gmail using OAuth2.
-
-    Filenames can be provided as arguments or will default to the values
-    specified in ``config.yaml``.
-    """
-    creds_filename = creds_filename or GMAIL_CLIENT_SECRET
-    token_filename = token_filename or GMAIL_TOKEN_FILE
-    creds = None
-    if os.path.exists(token_filename):
-        with open(token_filename, "rb") as t:
-            creds = pickle.load(t)
-    if not creds or not creds.valid:
-        if creds and creds.expired and creds.refresh_token:
-            creds.refresh(Request())
-        else:
-            flow = InstalledAppFlow.from_client_secrets_file(creds_filename, SCOPES)
-            creds = flow.run_local_server(port=0)
-        with open(token_filename, "wb") as t:
-            pickle.dump(creds, t)
-    return build("gmail", "v1", credentials=creds)
-
-def fetch_all_unread_messages(service):
-    unread, token = [], None
-    while True:
-        resp = service.users().messages().list(userId="me", q="is:unread", pageToken=token).execute()
-        unread.extend(resp.get("messages", []))
-        token = resp.get("nextPageToken")
-        if not token: break
-    return unread
-
-def create_base64_message(sender, to, subject, body):
-    from email.mime.text import MIMEText
-    msg = MIMEText(body)
-    msg["to"], msg["from"], msg["subject"] = to, sender, subject
-    return {"raw": base64.urlsafe_b64encode(msg.as_bytes()).decode()}
-
-def create_draft(service, user_id, msg_body, thread_id=None):
-    data = {"message": msg_body}
-    if thread_id: data["message"]["threadId"] = thread_id
-    return service.users().drafts().create(userId=user_id, body=data).execute()
-
-def thread_has_draft(service, thread_id):
-    data = service.users().threads().get(userId="me", id=thread_id).execute()
-    return any(
-        "DRAFT" in (m.get("labelIds") or []) for m in data.get("messages", [])
-    )
-
-def is_promotional_or_spam(message, body_text: str) -> bool:
-    """Return True if the message looks like a newsletter or spam."""
-    labels = set(message.get("labelIds", []))
-    if labels & PROMO_LABELS:
-        return True
-    headers = {h.get("name", "").lower(): h.get("value", "") for h in message.get("payload", {}).get("headers", [])}
-    if "list-unsubscribe" in headers or "list-id" in headers:
-        return True
-    if "unsubscribe" in body_text.lower():
-        return True
-    return False
-
-def critic_email(draft: str, original: str) -> dict:
-    """Self-grade a draft reply using GPT-4.1."""
-    client = OpenAI(api_key=OPENAI_API_KEY)
-    resp = client.chat.completions.create(
-        model=CLASSIFY_MODEL,
-        messages=[
-            {
-                "role": "system",
-                "content": (
-                    "Return ONLY JSON {\"score\":1-10,\"feedback\":\"...\"} "
-                    "rating on correctness, tone, length."
-                ),
-            },
-            {"role": "assistant", "content": draft},
-            {"role": "user", "content": f"Original email:\n\n{original}"},
-        ],
-    )
-    return json.loads(resp.choices[0].message.content)
-
-def classify_email(text: str) -> dict:
-    """Classify an email and return a dict with type and importance."""
-    client = OpenAI(api_key=OPENAI_API_KEY)
-    try:
-        response = client.chat.completions.create(
-            model=CLASSIFY_MODEL,
-            messages=[
-                {
-                    "role": "system",
-                    "content": (
-                        "Categorize the email as lead, customer, or other. Return ONLY JSON {\"type\":\"lead|customer|other\",\"importance\":1-10}. NO other text."
-                    ),
-                },
-                {"role": "user", "content": text},
-            ],
-            temperature=0,
-            max_tokens=CLASSIFY_MAX_TOKENS,
-        )
-        return json.loads(response.choices[0].message.content)
-    except Exception as e:
-        print(f"Error classifying email: {e}")
-        return {"type": "other", "importance": 0}
-
-
 
 def create_ticket(subject: str, sender: str, body: str):
     if TICKET_SYSTEM != "freescout":
